// types/nutrisi.d.ts
export interface NutrisiType{
    id: number
    title: string
    image: string
    definisi: string
    kebutuhan: string
    dianjurkan: string
    dibatasi: string
    dihindari: string
    definisiSound: string
    kebutuhanSound: string
    dianjurkanSound: string
    dibatasiSound: string
    dihindariSound: string
}